requirements.txt
# Core LLM and AI
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
websockets>=13.0

# Document Processing
PyPDF2>=3.0.0
pymupdf>=1.24.0
arxiv>=2.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
html2text>=2024.2.0
markdownify>=0.13.0

# GitHub and Git
PyGithub>=2.4.0
gitpython>=3.1.43

# Knowledge Graph
networkx>=3.4.0
rdflib>=7.0.0

# Data Processing
pydantic>=2.9.0
pandas>=2.2.0
numpy>=1.26.0

# Visualization
matplotlib>=3.9.0
plotly>=5.24.0
seaborn>=0.13.0

# Docker SDK
docker>=7.1.0

# HTTP and Async
httpx>=0.27.0
aiohttp>=3.10.0
aiofiles>=24.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.9.0
jinja2>=3.1.0
pyyaml>=6.0.0
tenacity>=9.0.0

# Report Generation
markdown>=3.7
weasyprint>=62.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
